Spin off MissionManagerStateMachine.h
